Move missing files to utils package
shared.h and sampling.h
