+++ apps/demo-utils.h
@@ -49,7 +49,7 @@
 #include <iostream>
 #include <string>
 
-#include <gr/shared.h>
+#include <gr/utils/shared.h>
 #include <gr/utils/logger.h>
 
 namespace gr {
