+++ src/gr/algorithms/congruentSetExplorationBase.hpp
@@ -11,15 +11,15 @@
 #include <omp.h>
 #endif
 
-#include "gr/shared.h"
-#include "gr/sampling.h"
+#include "gr/utils/shared.h"
+#include "gr/utils/sampling.h"
 #include "gr/accelerators/kdtree.h"
 #include "gr/utils/logger.h"
 
 #include "gr/algorithms/congruentSetExplorationBase.h"
 
 #ifdef TEST_GLOBAL_TIMINGS
-#   include "../utils/timer.h"
+#   include "gr/utils/timer.h"
 #endif
