+++ src/gr/algorithms/match4pcsBase.h
@@ -11,8 +11,8 @@
 #include <omp.h>
 #endif
 
-#include "gr/shared.h"
-#include "gr/sampling.h"
+#include "gr/utils/shared.h"
+#include "gr/utils/sampling.h"
 #include "gr/accelerators/kdtree.h"
 #include "gr/utils/logger.h"
 #include "gr/algorithms/congruentSetExplorationBase.h"
