+++ src/gr/algorithms/matchBase.h
@@ -50,8 +50,8 @@
 #include <omp.h>
 #endif
 
-#include "gr/shared.h"
-#include "gr/sampling.h"
+#include "gr/utils/shared.h"
+#include "gr/utils/sampling.h"
 #include "gr/accelerators/kdtree.h"
 #include "gr/utils/logger.h"
 #include "gr/utils/crtp.h"
