+++ src/gr/utils/shared.h
