+++ tests/testing.h
@@ -57,7 +57,7 @@
 #include <sstream>
 #include <ctime>
 
-#include "gr/shared.h"
+#include "gr/utils/shared.h"
 #include "gr/algorithms/matchBase.h"
 
 #define DEFAULT_REPEAT 10
